requirements.txt
# Core dependencies
lxml>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.0
